+++ HISTORY.rst
@@ -9,6 +9,8 @@ History
 - Drop support for Python 2 and Django 1.11, 2.0 and 2.1 (#155)
 - Template tag `bootstrap_field` now allows 3 values for `show_label`: `True`, `False` / `'sr-only'` and `'skip'`. In the case of `False` / `'sr-only'` the label is hidden but present for screen readers. When `show_label` is set to `'skip'`, the label is not generated at all.
 - Fix validation on input groups (#122)
+- No longer duplicate jquery in bootstrap4.html (#139, #140)
+- Apply form-check, form-check-label and form-check-input classes to appropriate tags for RadioSelect and CheckboxSelectMultiple widgets (#141)
 
 0.0.8 (2019-03-12)
 ++++++++++++++++++